+++ app/__init__.py
@@ -1,2 +1,2 @@
 """whisper-server FastAPI app"""
-__version__ = "0.1.1"
+__version__ = "0.1.2"